+++ pyproject.toml
@@ -30,7 +30,15 @@ dependencies = ["pandas>=2.0.0rc1"]
 repository = "https://github.com/thepolicylab/ZenTables.git"
 
 [project.optional-dependencies]
-dev = ["black[jupyter]", "ruff", "pytest", "pytest-asyncio", "mypy", "jupyterlab", "pre-commit"]
+dev = [
+    "black[jupyter]",
+    "ruff",
+    "pytest",
+    "mypy",
+    "jupyterlab",
+    "pre-commit",
+    "isort",
+]
 
 [tool.hatch.envs.dev]
 python = "3.9"
@@ -159,35 +167,6 @@ exclude = [".github", "docs", "src/notebook"]
 [tool.mypy]
 ignore_missing_imports = true
 python_version = "3.9"
-
-[tool.isort]
-known_pre_libs = "pandas._config"
-known_pre_core = [
-    "pandas._libs",
-    "pandas._typing",
-    "pandas.util._*",
-    "pandas.compat",
-    "pandas.errors",
-]
-known_dtypes = "pandas.core.dtypes"
-known_post_core = ["pandas.tseries", "pandas.io", "pandas.plotting"]
-sections = [
-    "FUTURE",
-    "STDLIB",
-    "THIRDPARTY",
-    "PRE_LIBS",
-    "PRE_CORE",
-    "DTYPES",
-    "FIRSTPARTY",
-    "POST_CORE",
-    "LOCALFOLDER",
-]
-profile = "black"
-combine_as_imports = true
-force_grid_wrap = 2
-force_sort_within_sections = true
-skip_glob = "env"
-skip = "pandas/__init__.py"
 
 [build-system]
 requires = ["hatchling"]
