+++ src/zentables/__init__.py
@@ -13,6 +13,7 @@ Typical usage examples::
     df.zen.pretty()
 """
 
-from .zentables import ZenTablesAccessor, set_options
+from .accessor import ZenTablesAccessor
+from .options import set_options
 
 __all__ = ["ZenTablesAccessor", "set_options"]
